Created cascading-expression sub-project to isolate all 'expression' operations based on Janino and it isolate the Janino dependency. A dependency to cascading-expression must be added to projects that depend on the isolated classes.
