+++ cascading-core/build.gradle
@@ -22,7 +22,6 @@ dependencies {
   compile group: 'org.jgrapht', name: 'jgrapht-core', version: '0.9.1'
   compile group: 'org.jgrapht', name: 'jgrapht-ext', version: '0.9.1'
   compile group: 'riffle', name: 'riffle', version: '1.0.0'
-  compile group: 'org.codehaus.janino', name: 'janino', version: '2.7.6'
 
   // this prevents a transitive dependency until it is required downstream
   provided group: 'org.slf4j', name: 'slf4j-api', version: '1.7.5'
