+++ cascading-expression/src/test/java/cascading/operation/expression/ScriptTest.java
